Implement MQTT Client Dependency Injection Framework
This commit introduces a dependency injection framework for MQTT client
selection, enabling runtime choice of MQTT client implementations while
maintaining backward compatibility.

Key Changes:
- Created mqtt_clients package with abstract interfaces
- Implemented factory pattern for client creation
- Moved existing MQTT classes to PahoMQTTPublisher/Subscriber
- Added --client-type CLI argument to both waldo.py and found-him.py
- Updated documentation with DI architecture details
- Maintained 100% backward compatibility (default: paho client)

Architecture:
- Abstract interfaces define standard MQTT operations
- Factory pattern enables extensible client selection
- Clean separation between client logic and implementation
- Ready for future client additions (e.g., NanoSDK)

Benefits:
- Runtime client selection without code changes
- Extensible design for future MQTT client support
- Better testability through interface abstractions
- Maintains existing functionality and performance
